Add files via upload
The data is related with direct marketing campaigns of a Portuguese banking institution. The marketing campaigns were based on phone calls. Often, more than one contact to the same client was required, in order to access if the product (bank term deposit) would be subscribed or not. This dataset was chosen in order to produce the best model for prediction of if the customer will make a deposit or not.